Make it work again
GitHub changed the DOM structure, so traversing for the URLs wasn't
working, nor was injecting the "Preview HTML" button.

I added a background script to offload the ajax requests because it was
getting Content Security Policy errors trying to run the ajax in the
content script. The background script is able to run the ajax requests
and it just sends the responses back to the content script.